+++ src/main/java/dev/chaws/automaticinventory/commands/AutoRefillCommand.java
@@ -1,21 +1,22 @@
 package dev.chaws.automaticinventory.commands;
 
 import dev.chaws.automaticinventory.*;
+import dev.chaws.automaticinventory.configuration.Features;
 import dev.chaws.automaticinventory.configuration.PlayerConfig;
 import dev.chaws.automaticinventory.messaging.Messages;
 import dev.chaws.automaticinventory.utilities.Chat;
-import dev.chaws.automaticinventory.utilities.TextMode;
+import dev.chaws.automaticinventory.utilities.Level;
 import org.bukkit.entity.Player;
 
 public class AutoRefillCommand implements IAutomaticInventoryCommand {
 	public boolean execute(Player player, PlayerConfig playerConfig, String[] args) {
 		if (!AutomaticInventory.hasPermission(Features.RefillStacks, player)) {
-			Chat.sendMessage(player, TextMode.Err, Messages.NoPermissionForFeature);
+			Chat.sendMessage(player, Level.Error, Messages.NoPermissionForFeature);
 			return true;
 		}
 
 		if (args.length < 1) {
-			Chat.sendMessage(player, TextMode.Instr, Messages.AutoRefillHelp);
+			Chat.sendMessage(player, Level.Instr, Messages.AutoRefillHelp);
 			return true;
 		}
 
@@ -24,18 +25,18 @@ public class AutoRefillCommand implements IAutomaticInventoryCommand {
 			playerConfig.setAutoRefillEnabled(!playerConfig.isAutoRefillEnabled());
 
 			if (playerConfig.isAutoRefillEnabled()) {
-				Chat.sendMessage(player, TextMode.Success, Messages.AutoRefillEnabled);
+				Chat.sendMessage(player, Level.Success, Messages.AutoRefillEnabled);
 			} else {
-				Chat.sendMessage(player, TextMode.Success, Messages.AutoRefillDisabled);
+				Chat.sendMessage(player, Level.Success, Messages.AutoRefillDisabled);
 			}
 		} else if (optionName.startsWith("enable")) {
 			playerConfig.setAutoRefillEnabled(true);
-			Chat.sendMessage(player, TextMode.Success, Messages.AutoRefillEnabled);
+			Chat.sendMessage(player, Level.Success, Messages.AutoRefillEnabled);
 		} else if (optionName.startsWith("disable")) {
 			playerConfig.setAutoRefillEnabled(false);
-			Chat.sendMessage(player, TextMode.Success, Messages.AutoRefillDisabled);
+			Chat.sendMessage(player, Level.Success, Messages.AutoRefillDisabled);
 		} else {
-			Chat.sendMessage(player, TextMode.Err, Messages.AutoRefillHelp);
+			Chat.sendMessage(player, Level.Error, Messages.AutoRefillHelp);
 			return true;
 		}
