+++ src/main/java/dev/chaws/automaticinventory/commands/QuickDepositCommand.java
@@ -1,6 +1,7 @@
 package dev.chaws.automaticinventory.commands;
 
 import dev.chaws.automaticinventory.*;
+import dev.chaws.automaticinventory.configuration.Features;
 import dev.chaws.automaticinventory.configuration.PlayerConfig;
 import dev.chaws.automaticinventory.messaging.Messages;
 import dev.chaws.automaticinventory.utilities.*;
@@ -11,12 +12,12 @@ import static dev.chaws.automaticinventory.AutomaticInventory.hasPermission;
 public class QuickDepositCommand implements IAutomaticInventoryCommand {
 	public boolean execute(Player player, PlayerConfig playerConfig, String[] args) {
 		if (!hasPermission(Features.QuickDeposit, player)) {
-			Chat.sendMessage(player, TextMode.Err, Messages.NoPermissionForFeature);
+			Chat.sendMessage(player, Level.Error, Messages.NoPermissionForFeature);
 			return true;
 		}
 
 		if (args.length < 1) {
-			Chat.sendMessage(player, TextMode.Instr, Messages.QuickDepositHelp);
+			Chat.sendMessage(player, Level.Instr, Messages.QuickDepositHelp);
 			return true;
 		}
 
@@ -25,18 +26,18 @@ public class QuickDepositCommand implements IAutomaticInventoryCommand {
 			playerConfig.setQuickDepositEnabled(!playerConfig.isQuickDepositEnabled());
 
 			if (playerConfig.isQuickDepositEnabled()) {
-				Chat.sendMessage(player, TextMode.Success, Messages.QuickDepositEnabled);
+				Chat.sendMessage(player, Level.Success, Messages.QuickDepositEnabled);
 			} else {
-				Chat.sendMessage(player, TextMode.Success, Messages.QuickDepositDisabled);
+				Chat.sendMessage(player, Level.Success, Messages.QuickDepositDisabled);
 			}
 		} else if (optionName.startsWith("enable")) {
 			playerConfig.setQuickDepositEnabled(true);
-			Chat.sendMessage(player, TextMode.Success, Messages.QuickDepositEnabled);
+			Chat.sendMessage(player, Level.Success, Messages.QuickDepositEnabled);
 		} else if (optionName.startsWith("disable")) {
 			playerConfig.setQuickDepositEnabled(false);
-			Chat.sendMessage(player, TextMode.Success, Messages.QuickDepositDisabled);
+			Chat.sendMessage(player, Level.Success, Messages.QuickDepositDisabled);
 		} else {
-			Chat.sendMessage(player, TextMode.Err, Messages.QuickDepositHelp);
+			Chat.sendMessage(player, Level.Error, Messages.QuickDepositHelp);
 			return true;
 		}
