+++ src/main/java/dev/chaws/automaticinventory/common/DepositRecord.java
@@ -1,7 +1,7 @@
-package dev.chaws.automaticinventory;
+package dev.chaws.automaticinventory.common;
 
 public class DepositRecord {
-	DepositRecord() { }
+	public DepositRecord() { }
 
 	public int totalItems = 0;
 	public boolean destinationFull = false;
