+++ src/main/java/dev/chaws/automaticinventory/messaging/Messages.java
@@ -1,5 +1,3 @@
-//Copyright 2015 Ryan Hamshire
-
 package dev.chaws.automaticinventory.messaging;
 
 public enum Messages {
