+++ src/main/java/dev/chaws/automaticinventory/utilities/ItemUtilities.java
@@ -1,5 +1,6 @@
 package dev.chaws.automaticinventory.utilities;
 
+import org.bukkit.enchantments.Enchantment;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.PotionMeta;
 
@@ -24,5 +25,36 @@ public class ItemUtilities {
 		}
 
 		return signature;
+	}
+
+	public static boolean itemsAreSimilar(ItemStack a, ItemStack b) {
+		if (a.getType() == b.getType()) {
+			if (a.containsEnchantment(Enchantment.LOOT_BONUS_BLOCKS) || a.containsEnchantment(Enchantment.SILK_TOUCH) || a.containsEnchantment(Enchantment.LOOT_BONUS_MOBS)) {
+				return false;
+			}
+
+			//a will _not_ have itemMeta if it is a vanilla tool with no damage.
+//            if(a.hasItemMeta() != b.hasItemMeta()) return false;
+//
+//            //compare metadata
+//            if(a.hasItemMeta())
+//            {
+//                if(!b.hasItemMeta()) return false;
+//
+//                ItemMeta meta1 = a.getItemMeta();
+//                ItemMeta meta2 = b.getItemMeta();
+//
+//                //compare names
+//                if(meta1.hasDisplayName())
+//                {
+//                    if(!meta2.hasDisplayName()) return false;
+//                    return meta1.getDisplayName().equals(meta2.getDisplayName());
+//                }
+//            }
+
+			return true;
+		}
+
+		return false;
 	}
 }
