+++ src/main/java/dev/chaws/automaticinventory/utilities/Level.java
@@ -2,10 +2,10 @@ package dev.chaws.automaticinventory.utilities;
 
 import org.bukkit.ChatColor;
 
-public class TextMode {
+public class Level {
 	public final static ChatColor Info = ChatColor.AQUA;
 	public final static ChatColor Instr = ChatColor.YELLOW;
 	public final static ChatColor Warn = ChatColor.GOLD;
-	public final static ChatColor Err = ChatColor.RED;
+	public final static ChatColor Error = ChatColor.RED;
 	public final static ChatColor Success = ChatColor.GREEN;
 }
